+++ reproc/test/argv.cpp
@@ -1,10 +1,9 @@
 #include <doctest/doctest.h>
 #include <reproc/reproc.h>
 
+#include <algorithm>
 #include <array>
 #include <string>
-#include <algorithm>
-#include <iostream>
 
 TEST_SUITE("reproc")
 {
@@ -12,28 +11,29 @@ TEST_SUITE("reproc")
   {
     reproc_t process;
 
+    static constexpr unsigned int ARGV_SIZE = 4;
     std::array<const char *, 4> argv = { "reproc/resources/argv", "argument 1",
                                          "argument 2", nullptr };
 
-    REPROC_ERROR error = reproc_start(&process, argv.size() - 1, argv.data(),
+    REPROC_ERROR error = reproc_start(&process, ARGV_SIZE - 1, argv.data(),
                                       nullptr);
     REQUIRE(!error);
 
     std::string output;
-    std::array<uint8_t, 1024> buffer = {};
+
+    static constexpr unsigned int BUFFER_SIZE = 4;
+    std::array<uint8_t, BUFFER_SIZE> buffer = {};
 
     while (true) {
       unsigned int bytes_read = 0;
       error = reproc_read(&process, REPROC_STREAM_OUT, buffer.data(),
-                          buffer.size(), &bytes_read);
+                          BUFFER_SIZE, &bytes_read);
       if (error != REPROC_SUCCESS) {
         break;
       }
 
       output.append(reinterpret_cast<const char *>(buffer.data()), bytes_read);
     }
-
-    std::cout << output << std::endl;
 
     REQUIRE(error == REPROC_ERROR_STREAM_CLOSED);
     REQUIRE(std::count(output.begin(), output.end(), '\n') == argv.size() - 1);